@@ -1,38 +1,22 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "npm run development -- --watch",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "production-with-no-cross-env": "NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "dev": "vite",
+        "build": "vite build"
     },
     "devDependencies": {
-        "bootstrap": "^4.5.0",
-        "cross-env": "^5.2.1",
-        "jquery": "^3.6.0",
-        "laravel-mix": "^4.1.4",
-        "lodash": "^4.17.21",
+        "@vitejs/plugin-vue": "^3.2.0",
+        "axios": "^1.1.2",
+        "laravel-vite-plugin": "^0.6.0",
+        "lodash": "^4.17.19",
         "popper.js": "^1.16.1",
-        "resolve-url-loader": "^2.3.1",
-        "sass": "^1.32.13",
-        "sass-loader": "^7.3.1",
-        "vue": "^2.6.14",
-        "vue-template-compiler": "^2.6.14"
-    },
-    "dependencies": {
-        "acorn-dynamic-import": "^4.0.0",
-        "axios": "^0.19.2",
-        "css-unit-converter": "^1.1.2",
-        "css-url-regex": "^3.3.1",
-        "for-own": "^1.0.0",
-        "lodash.tail": "^4.1.1",
-        "mixin-object": "^3.0.0",
-        "select2": "^4.0.13",
-        "v-tooltip": "^2.1.3",
-        "vue-d3-charts": "^0.2.8"
+        "postcss": "^8.1.14",
+        "sass": "^1.56.0",
+        "floating-vue": "^2.0.0-beta.20",
+        "vite": "^3.0.0",
+        "vue": "^3.2.36",
+        "bootstrap": "^4.5.0",
+        "vue-loader": "^17.0.0",
+        "vue-select" : "beta"
     }
 }
