@@ -1,15 +1,21 @@
 
+import GroupEnrollView from "./views/GroupEnrollView";
+import MergeUserView from "./views/MergeUserView";
+import DiplomaView from "./views/DiplomaView";
+import NoDiplomaView from "./views/NoDiplomaView";
+import KpasEmbedView from "./views/KpasEmbedView";
+import {createApp} from "vue/dist/vue.esm-bundler";
+import './bootstrap';
+import FloatingVue from 'floating-vue';
+import vSelect from "vue-select";
+
+//import jqueryExports from "jquery";
+//import 'select2';
 /**
  * First we will load all of this project's JavaScript dependencies which
  * includes Vue and other libraries. It is a great starting point when
  * building robust, powerful web applications using Vue and Laravel.
  */
-
-try {
-    window.$ = window.jQuery = require('jquery');
-    window.Vue = require('vue');
-    require('select2');
-} catch (e) {}
 
 /**
  * The following block of code may be used to automatically register your
@@ -21,24 +27,22 @@ try {
 
 // const files = require.context('./', true, /\.vue$/i)
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
-
-Vue.component('group-enroll-view', require('./views/GroupEnrollView').default);
-Vue.component('merge-user-view', require('./views/MergeUserView').default);
-Vue.component('diploma-view', require('./views/DiplomaView').default);
-Vue.component('no-diploma-view', require('./views/NoDiplomaView').default);
-Vue.component('kpas-embed-view', require('./views/KpasEmbedView').default);
-Vue.component('statistics-view', require('./views/StatisticsView').default);
 /**
  * Next, we will create a fresh Vue application instance and attach it to
  * the page. Then, you may begin adding components to this application
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-const app = new Vue({
-    el: '#app',
-    computed: {
-      window() {
-        return window;
-      }
-    }
-});
+//window.$ = jqueryExports.default;
+
+const app = createApp({});
+
+app.use(FloatingVue);
+app.component('group-enroll-view', GroupEnrollView);
+app.component('merge-user-view', MergeUserView);
+app.component('diploma-view', DiplomaView);
+app.component('no-diploma-view', NoDiplomaView);
+app.component('kpas-embed-view', KpasEmbedView);
+app.component("v-select", vSelect);
+
+app.mount("#app")
