@@ -1,15 +1,15 @@
 <template>
-    <div>
-        <span v-html="information"></span>
-    </div>
+  <div>
+      <span v-html="information"></span>
+  </div>
 </template>
 
 <script>
-  export default {
-    name: "CurrentRole",
-    props: {
-        isPrincipal: Boolean,
-        information: String
-    }
+export default {
+  name: "CurrentRole",
+  props: {
+      isPrincipal: Boolean,
+      information: String
   }
+}
 </script>
