@@ -1,9 +1,9 @@
 <template>
   <div>
-      <input type="radio" id="radioSkoleleder" name="role" v-bind:value="true" v-model="wantToBePrincipal">
+      <input type="radio" id="radioSkoleleder" name="role" :value=true v-model="wantToBePrincipal" @input="$emit('update:modelValue', true)">
       <label for="radioSkoleleder">{{leaderDescription}}</label>
       <br>
-      <input type="radio" id="deltager" name="role" v-bind:value="false" v-model="wantToBePrincipal">
+      <input type="radio" id="deltager" name="role" :value=false v-model="wantToBePrincipal" @input="$emit('update:modelValue', false)">
       <label for="deltager">{{participantDescription}}</label>
     <div v-if="wantToBePrincipal && institutionType" class="alert alert-info">{{principalWarning}}
     </div>
@@ -11,14 +11,15 @@
     </div>
 </template>
 
-<script>
+<script lang="js">
   export default {
     name: "RoleSelector",
     props: {
         isPrincipal: Boolean,
         institutionType: String,
         leaderDescription: String,
-        participantDescription: String
+        participantDescription: String,
+        modelValue: Boolean
     },
     data() {
       return {
@@ -38,14 +39,12 @@
     created() {
       this.wantToBePrincipal = this.isPrincipal;
     },
+    emits: ['update:value'],
     watch: {
       isPrincipal(value) {
         this.wantToBePrincipal = value;
         this.$parent.iframeresize();
       },
-      wantToBePrincipal(value) {
-        this.$emit('input', value);
-      },
-    },
+    }
   }
 </script>
